Made karma go up to 1000 and is a double.

+++ src/main/java/com/blockmovers/plugins/pvpkarma/Karma.java
@@ -54,11 +54,11 @@ public class Karma {
         }
     }
     
-    public int getKarma(String p) {
-        return this.karma.getInt("karma." + p, 0);
+    public double getKarma(String p) {
+        return this.karma.getDouble("karma." + p, 0);
     }
     
-    public void setKarma(String p, int k) {
+    public void setKarma(String p, double k) {
         if (k > 1000) {
             k = 1000;
         }
